+++ BubbleTextField/BubbleTextField.swift
@@ -19,6 +19,7 @@ open class BubbleTextField: NSTextField {
             bubbleTextFieldCell?.isContinuousSpellCheckingEnabled = isContinuousSpellCheckingEnabled
         }
     }
+    public weak var textViewDelegate: NSTextViewDelegate?
 
     public override init(frame frameRect: NSRect) {
         super.init(frame: frameRect)
@@ -62,6 +63,21 @@ open class BubbleTextField: NSTextField {
         needsUpdateConstraints = true
     }
 
+}
+
+extension BubbleTextField: NSTextViewDelegate {
+    @available(OSX 10.12.2, *)
+    public func textView(_ textView: NSTextView, shouldSelectCandidateAt index: Int) -> Bool {
+        return textViewDelegate?.textView?(textView, shouldSelectCandidateAt: index) ?? true
+    }
+
+    public func textView(_ textView: NSTextView, shouldChangeTextIn affectedCharRange: NSRange, replacementString: String?) -> Bool {
+        return textViewDelegate?.textView?(textView, shouldChangeTextIn: affectedCharRange, replacementString: replacementString) ?? true
+    }
+
+    public func textView(_ view: NSTextView, didCheckTextIn range: NSRange, types checkingTypes: NSTextCheckingTypes, options: [NSSpellChecker.OptionKey : Any] = [:], results: [NSTextCheckingResult], orthography: NSOrthography, wordCount: Int) -> [NSTextCheckingResult] {
+        return textViewDelegate?.textView?(view, didCheckTextIn: range, types: checkingTypes, options: options, results: results, orthography: orthography, wordCount: wordCount) ?? results
+    }
 }
 
 private class BubbleTextFieldCell: NSTextFieldCell {